Simplify structure of knapsack example
