+++ optiframe/workflow_engine/task.py
@@ -11,9 +11,9 @@ class Task(abc.ABC, Generic[T]):
     The task can generate base_data and return it from the `execute` function.
     This base_data will be made available to the other tasks of the workflow.
 
-    The task can also depend on base_data by other tasks by specifying dependencies in the constructor.
+    The task can also depend on data by other tasks by specifying dependencies in the constructor.
     The type annotation of the constructor parameters determine which dependency will be injected.
-    If the base_data is generated by a task in the same step, the task will be executed before this one.
+    If the data is generated by a task in the same step, the task will be executed before this one.
     """
 
     @abc.abstractmethod
